requirements.txt
requests
spotipy
python-dotenv
honcho